+++ src/data.ts
@@ -1,0 +1,7 @@
+import { RegisteredAttribute } from "./interfaces/RegisteredAttribute";
+import { RegisteredModel } from "./interfaces/RegisteredModel";
+
+export const $registeredModels: RegisteredModel[] = [];
+export const $registeredAttributes: RegisteredAttribute[] = [];
+export const $registeredRelationships: RegisteredAttribute[] = [];
+