+++ src/decorators/JSONAPI.ts
@@ -1,9 +1,9 @@
+import { $registeredModels } from "../data";
 import { Model } from "../Model";
-import { Parser } from "../Parser";
 
 export function JSONAPI(type: string) {
   return function _JSONAPI<T extends typeof Model>(constructor: T) {
-    Parser.$registeredModels.push({
+    $registeredModels.push({
       klass: constructor,
       type,
     });